+++ packages/terra-toggle/src/terra-dev-site/doc/example/AnimatedToggle.jsx
@@ -19,8 +19,11 @@ class ToggleDefault extends React.Component {
     const { isOpen } = this.state;
     return (
       <div>
-        <IconInformation onClick={this.handleOnClick} />
-        <Toggle isOpen={isOpen} isAnimated>
+        <IconInformation onClick={this.handleOnClick} aria-expanded={isOpen} aria-controls="toggle" />
+        {/**
+        * The aria-expanded state is used on the triggering component to indicate the contents are collapsible, and whether a region is currently expanded or collapsed
+        */}
+        <Toggle id="toggle" isOpen={isOpen} isAnimated>
           <p>
 Lorem ipsum dolor sit amet,
             <a href="#test">consectetur</a>