+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+1.1.0 - (August 14, 2019)
+-------------------------
+### Changed
+* Minor dependency updates
+
 1.0.0 - (July 30, 2019)
 ------------------
 * Initial stable release