+++ packages/terra-markdown/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Added
 * Added package.json import to doc page
 
+2.32.0 - (August 14, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.31.0 - (July 30, 2019)
 ------------------
 ### Changed